fix cards.js and home.js
